Authentication

+++ app/api/views.py
@@ -9,23 +9,58 @@ from sqlalchemy.sql import func, column, text
 from sqlalchemy.sql.expression import cast, true
 from . import api
 from flask import request, jsonify
+from flask_jwt_simple import (
+    jwt_required, create_jwt, get_jwt_identity
+)
 from .forms import newShelterForm
 from ..models import db, Shelter, Count, Log
 
 tz = os.environ['PEND_TZ'] 
+
+##############
+#### AUTH ####
+##############
+
+@api.route('/admin_login/', methods = ['POST'])
+@cross_origin()
+def login():
+    if not request.is_json:
+        return jsonify({"msg": "Missing JSON in request"}), 400
+
+    params = request.get_json()
+    user = params.get('user', None)
+    password = params.get('password', None)
+    if not user:
+        return jsonify({"msg": "Missing username parameter"}), 400
+    if not password:
+        return jsonify({"msg": "Missing password parameter"}), 400
+
+    if user != os.environ['ADMIN_USER'] or password != os.environ['ADMIN_PW']:
+        return jsonify({"msg": "Bad username or password"}), 401
+
+    ret = {'jwt': create_jwt(identity=user)}
+    return jsonify(ret), 200
 
 ##################
 #### SHELTERS ####
 ##################
 @api.route('/shelters/', methods = ['GET', 'POST'])
 @cross_origin()
+@jwt_required
 def get_shelters():
+    if get_jwt_identity() != os.environ['ADMIN_USER']:
+        return jsonify(msg="Permission denied"), 403
+
     shelters = Shelter.query.order_by('name').all()
     return jsonify([s.toDict() for s in shelters])
 
 @api.route('/add_shelters/', methods = ['GET', 'POST'])
 @cross_origin()
+@jwt_required
 def add_shelter():
+    if get_jwt_identity() != os.environ['ADMIN_USER']:
+        return jsonify(msg="Permission denied"), 403
+
     form = newShelterForm()
     if form.validate_on_submit():
         shelter = {}
@@ -46,14 +81,22 @@ def add_shelter():
 
 @api.route('/delete_shelter/<shelter_id>', methods = ['GET'])
 @cross_origin()
+@jwt_required
 def delete_shelter(shelter_id):
+    if get_jwt_identity() != os.environ['ADMIN_USER']:
+        return jsonify(msg="Permission denied"), 403
+
     Shelter.query.filter_by(id=shelter_id).delete()
     db.session.commit()
     return jsonify({"result":"success"})
 
 @api.route('/update_shelter/', methods=['POST'])
 @cross_origin()
+@jwt_required
 def update_shelter():
+    if get_jwt_identity() != os.environ['ADMIN_USER']:
+        return jsonify(msg="Permission denied"), 403
+
     form = newShelterForm()
     shelter = {}
     shelter['id']       = form.id.data
@@ -81,10 +124,13 @@ def update_shelter():
 @api.route('/counts/', methods=['GET'], defaults = {'daysback': 0})
 @api.route('/counts/<daysback>', methods=['GET'])
 @cross_origin()
+@jwt_required
 def counts(daysback):
     ''' Results the lastest counts per shelter '''
+    if get_jwt_identity() != os.environ['ADMIN_USER']:
+        return jsonify(msg="Permission denied"), 403
+
     today = pendulum.today(tz).subtract(days = int(daysback))
-    print( today.isoformat(' '))
 
     count_calls = db.session.query(Count.shelter_id.label("call_shelterID"), Count.bedcount, Count.day, Count.time)\
                   .filter(Count.day == today.isoformat(' '))\
@@ -101,8 +147,12 @@ def counts(daysback):
 @api.route('/counthistory/', methods=['GET'], defaults = {'page': 0})
 @api.route('/counthistory/<page>/', methods=['GET'])
 @cross_origin()
+@jwt_required
 def counthistory(page):
     ''' Final count history for all shelters. Used for chart showing counts over time '''
+    if get_jwt_identity() != os.environ['ADMIN_USER']:
+        return jsonify(msg="Permission denied"), 403
+
     pagesize = 14 # days
     daysback = int(page) * pagesize + pagesize - 1
 
@@ -127,13 +177,16 @@ def counthistory(page):
 @api.route('/logs/<shelterid>/', methods=['GET'])
 @api.route('/logs/<shelterid>/<page>/', methods=['GET'])
 @cross_origin()
+@jwt_required
 def logs(shelterid, page=0):
     '''Provives a list of logs for a particular shelter'''
+    if get_jwt_identity() != os.environ['ADMIN_USER']:
+        return jsonify(msg="Permission denied"), 403
+
     pagesize = 15 #records
     offset = pagesize * int(page)
     shelter = Shelter.query.get_or_404(shelterid)
     total_calls = db.session.query(func.count(Log.id)).filter_by(shelter_id=shelterid).scalar()
-    print(total_calls)
     logs = db.session.query(Log)\
             .filter_by(shelter_id = shelterid)\
             .order_by(Log.time.desc())\
